+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/home/durandhil/test_tech/romain_test_tech_on_rewind/src/index.jsx":"1","/home/durandhil/test_tech/romain_test_tech_on_rewind/src/Router.jsx":"2","/home/durandhil/test_tech/romain_test_tech_on_rewind/src/Components/NavBar.jsx":"3","/home/durandhil/test_tech/romain_test_tech_on_rewind/src/Components/Query.js":"4","/home/durandhil/test_tech/romain_test_tech_on_rewind/src/Layout/Layout.jsx":"5","/home/durandhil/test_tech/romain_test_tech_on_rewind/src/Components/Routes.js":"6","/home/durandhil/test_tech/romain_test_tech_on_rewind/src/Pages/SingleVideoDisplay.jsx":"7","/home/durandhil/test_tech/romain_test_tech_on_rewind/src/Pages/ThumbnailsDiplay.jsx":"8"},{"size":1116,"mtime":1611277245120,"results":"9","hashOfConfig":"10"},{"size":234,"mtime":1611438682530,"results":"11","hashOfConfig":"10"},{"size":560,"mtime":1611440492640,"results":"12","hashOfConfig":"10"},{"size":613,"mtime":1611428343210,"results":"13","hashOfConfig":"10"},{"size":582,"mtime":1611440636960,"results":"14","hashOfConfig":"10"},{"size":729,"mtime":1611441408700,"results":"15","hashOfConfig":"10"},{"size":1033,"mtime":1611440337450,"results":"16","hashOfConfig":"10"},{"size":2673,"mtime":1611441296220,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},"1uxae5p",{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/durandhil/test_tech/romain_test_tech_on_rewind/src/index.jsx",[],["35","36"],"/home/durandhil/test_tech/romain_test_tech_on_rewind/src/Router.jsx",[],"/home/durandhil/test_tech/romain_test_tech_on_rewind/src/Components/NavBar.jsx",[],"/home/durandhil/test_tech/romain_test_tech_on_rewind/src/Components/Query.js",[],"/home/durandhil/test_tech/romain_test_tech_on_rewind/src/Layout/Layout.jsx",[],"/home/durandhil/test_tech/romain_test_tech_on_rewind/src/Components/Routes.js",[],"/home/durandhil/test_tech/romain_test_tech_on_rewind/src/Pages/SingleVideoDisplay.jsx",[],"/home/durandhil/test_tech/romain_test_tech_on_rewind/src/Pages/ThumbnailsDiplay.jsx",[],{"ruleId":"37","replacedBy":"38"},{"ruleId":"39","replacedBy":"40"},"no-native-reassign",["41"],"no-negated-in-lhs",["42"],"no-global-assign","no-unsafe-negation"]
+[{"/home/durandhil/test_tech/romain_test_tech_on_rewind/src/index.jsx":"1","/home/durandhil/test_tech/romain_test_tech_on_rewind/src/Router.jsx":"2","/home/durandhil/test_tech/romain_test_tech_on_rewind/src/Layout.jsx":"3","/home/durandhil/test_tech/romain_test_tech_on_rewind/src/Components/NavBar.jsx":"4","/home/durandhil/test_tech/romain_test_tech_on_rewind/src/Components/ThumbnailsDiplay.jsx":"5","/home/durandhil/test_tech/romain_test_tech_on_rewind/src/Components/Query.js":"6"},{"size":1116,"mtime":1611277245120,"results":"7","hashOfConfig":"8"},{"size":227,"mtime":1611183526240,"results":"9","hashOfConfig":"8"},{"size":993,"mtime":1611420101320,"results":"10","hashOfConfig":"8"},{"size":588,"mtime":1611420101320,"results":"11","hashOfConfig":"8"},{"size":2170,"mtime":1611420101320,"results":"12","hashOfConfig":"8"},{"size":428,"mtime":1611420101320,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},"1uxae5p",{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},"/home/durandhil/test_tech/romain_test_tech_on_rewind/src/index.jsx",[],["27","28"],"/home/durandhil/test_tech/romain_test_tech_on_rewind/src/Router.jsx",[],"/home/durandhil/test_tech/romain_test_tech_on_rewind/src/Layout.jsx",[],"/home/durandhil/test_tech/romain_test_tech_on_rewind/src/Components/NavBar.jsx",[],"/home/durandhil/test_tech/romain_test_tech_on_rewind/src/Components/ThumbnailsDiplay.jsx",[],"/home/durandhil/test_tech/romain_test_tech_on_rewind/src/Components/Query.js",[],{"ruleId":"29","replacedBy":"30"},{"ruleId":"31","replacedBy":"32"},"no-native-reassign",["33"],"no-negated-in-lhs",["34"],"no-global-assign","no-unsafe-negation"]
