+++ package.json
@@ -4,28 +4,20 @@
   "private": true,
   "dependencies": {
     "@apollo/client": "^3.3.7",
-    "@babel/preset-env": "^7.12.11",
-    "@babel/preset-react": "^7.12.10",
     "@material-ui/core": "^4.11.2",
     "@testing-library/jest-dom": "^5.11.9",
     "@testing-library/react": "^11.2.3",
     "@testing-library/user-event": "^12.6.0",
-    "babel-jest": "^26.6.3",
-    "eslint": "^7.18.0",
     "graphql": "^15.4.0",
-    "jest": "^26.6.0",
-    "prettier": "^2.2.1",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.1",
-    "react-test-renderer": "^17.0.1",
     "web-vitals": "^0.2.4"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
-    "test:jest": "jest",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
