+++ src/App.test.js
@@ -1,8 +1,8 @@
-import { render, screen } from "@testing-library/react";
-import Layout from "./Layout/Layout";
+import { render, screen } from '@testing-library/react';
+import App from './App';
 
-test("renders learn react link", () => {
-  render(<Layout />);
+test('renders learn react link', () => {
+  render(<App />);
   const linkElement = screen.getByText(/learn react/i);
   expect(linkElement).toBeInTheDocument();
 });
