+++ src/Components/Query.js
@@ -24,17 +24,3 @@ export const ALLVIDEOS = gql`
     }
   }
 `;
-
-export const VIDEO = gql`
-  query Getvideo($id: ID!) {
-    video(id: $id) {
-      name
-      poster
-      Tags {
-        name
-        tagType
-        placeholders
-      }
-    }
-  }
-`;
