+++ src/Router.jsx
@@ -1,5 +1,5 @@
 import { BrowserRouter, Route } from "react-router-dom";
-import Layout from "./Layout/Layout";
+import Layout from "./Layout";
 
 export default function Router() {
   return (
