+++ src/Styles/NavBar.css
@@ -9,7 +9,6 @@
 
 .logo {
   margin-left: 15%;
-  border-radius: 10px;
 }
 
 .navMenu {
