--- src/Styles/Thumbnails.css
@@ -1,124 +1,0 @@
-/* SINGLE VIDEO CSS */
-
-.singleVideoCard {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 80vw;
-  min-height: 35vh;
-  max-height: 75vh;
-  padding: 1em;
-  margin: 0 1em 1em 1em;
-  background-color: gray;
-  border-radius: 20px;
-  -webkit-box-shadow: -20px 0px 43px -7px #64adf7, 20px 0px 43px -7px #64adf7,
-    5px 5px 8px 5px rgba(0, 0, 0, 0);
-  box-shadow: -20px 0px 43px -7px rgb(100, 173, 247), 20px 0px 43px -7px #64adf7,
-    5px 5px 8px 5px rgba(0, 0, 0, 0);
-}
-
-.singleVideoPoster {
-  width: auto;
-  max-width: 60vw;
-  max-height: 65vh;
-  border-radius: 20px;
-}
-
-/* THUMBNAILS & SINGLE VIDEO COMMON CSS*/
-.title {
-  color: #fff;
-}
-
-.container {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: space-around;
-}
-
-.card {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 40%;
-  max-height: 30%;
-  padding: 1em;
-  margin: 1em;
-  background-color: gray;
-  border-radius: 20px;
-  -webkit-box-shadow: -20px 0px 43px -7px #fff, 20px 0px 43px -7px #fff,
-    15px 5px 38px 5px rgba(0, 0, 0, 0);
-  box-shadow: -20px 0px 43px -7px #fff, 20px 0px 43px -7px #fff,
-    15px 5px 38px 5px rgba(0, 0, 0, 0);
-}
-
-.card:hover {
-  scale: 1.009;
-  -webkit-box-shadow: -20px 0px 43px -7px #64adf7, 20px 0px 43px -7px #64adf7,
-    5px 5px 8px 5px rgba(0, 0, 0, 0);
-  box-shadow: -20px 0px 43px -7px #64adf7, 20px 0px 43px -7px #64adf7,
-    5px 5px 8px 5px rgba(0, 0, 0, 0);
-}
-
-.videoName {
-  margin: 0;
-  word-break: break-all;
-}
-
-.poster {
-  width: 100%;
-  border-radius: 20px;
-}
-
-.tags {
-  display: flex;
-  align-items: center;
-  flex-wrap: wrap;
-  height: 1em;
-}
-
-.tags h4 {
-  align-self: flex-start;
-  height: 1em;
-  margin: 0;
-}
-
-.tags p {
-  margin: 0;
-  height: 1em;
-}
-
-button {
-  height: 50px;
-  width: 150px;
-  border-radius: 15px;
-  background-color: gray;
-  margin: 10px 0 10px 10px;
-}
-
-button:hover {
-  background-color: rgb(37, 100, 163);
-}
-
-@media screen and (max-width: 1080px) {
-  .card {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    width: 80vw;
-    min-height: 35vh;
-    padding: 1em;
-    margin: 1em;
-    background-color: gray;
-    border-radius: 20px;
-  }
-
-  .poster {
-    max-width: 100%;
-  }
-
-  /* SINGLE VIDEO MEDIA QUERY */
-
-  .singleVideoPoster {
-    max-width: 100%;
-  }
-}
