+++ src/Styles/thumbnails.css
@@ -1,0 +1,85 @@
+.container {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: space-around;
+}
+
+.thumbnail {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 40%;
+  max-height: 30%;
+  padding: 1em;
+  margin: 1em;
+  background-color: gray;
+  border-radius: 20px;
+  -webkit-box-shadow: -10px 0px 13px -7px #fff, 10px 0px 13px -7px #fff,
+    5px 5px 8px 5px rgba(0, 0, 0, 0);
+  box-shadow: -10px 0px 13px -7px #fff, 10px 0px 13px -7px #fff,
+    5px 5px 8px 5px rgba(0, 0, 0, 0);
+}
+
+.thumbnail:hover {
+  scale: 1.009;
+  -webkit-box-shadow: -10px 0px 13px -7px rgb(37, 100, 163),
+    10px 0px 13px -7px rgb(37, 100, 163), 5px 5px 8px 5px #000;
+  box-shadow: -10px 0px 13px -7px rgb(37, 100, 163),
+    10px 0px 13px -7px rgb(37, 100, 163), 5px 5px 8px 5px#000;
+}
+
+.thumbnail h3 {
+  margin: 0;
+  word-break: break-all;
+}
+
+.thumbnail img {
+  width: 100%;
+  border-radius: 20px;
+}
+
+.tags {
+  display: flex;
+  align-items: center;
+  flex-wrap: wrap;
+  height: 1em;
+}
+
+.tags h4 {
+  align-self: flex-start;
+  height: 1em;
+  margin: 0;
+}
+
+.tags p {
+  margin: 0;
+  height: 1em;
+}
+
+button {
+  height: 50px;
+  width: 150px;
+  border-radius: 15px;
+  background-color: gray;
+  margin: 10px 0 10px 10px;
+}
+
+button:hover {
+  background-color: rgb(37, 100, 163);
+}
+
+@media screen and (max-width: 1080px) {
+  .thumbnail {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    width: 80vw;
+    padding: 1em;
+    margin: 1em;
+    background-color: gray;
+  }
+
+  .thumbnail img {
+    max-width: 100%;
+  }
+}
